pyproject.toml
[project]
name = "rest2mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "Gonzalo Ayuso",email = "gonzalo123@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]",
    "flask",
    "requests",
    "python-dotenv",
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
pythonPath = "src"
include = ["src"]
typeCheckingMode = "basic"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest",
]
